# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.User:
    properties:
      acedemic_year:
        type: integer
      blue_badge:
        type: boolean
      created_at:
        type: string
      department_id:
        type: string
      email:
        type: string
      follow_count:
        type: integer
      google_id:
        type: string
      id:
        type: string
      is_complete:
        type: boolean
      is_teacher:
        type: boolean
      is_verified:
        type: boolean
      name:
        type: string
      password:
        type: string
      profile_image_url:
        type: string
      role:
        type: string
      school_id:
        type: string
      university_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: IKnow Learning Platform API
  title: IKnow API
  version: "1.0"
paths:
  /api/auth/email/forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password email
      parameters:
      - description: User Email
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - Auth
  /api/auth/email/login:
    post:
      consumes:
      - application/json
      description: Authenticate user using email and password
      parameters:
      - description: User Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Returns JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email
      tags:
      - Auth
  /api/auth/email/register:
    post:
      consumes:
      - application/json
      description: Create a new account and send verification email
      parameters:
      - description: User Registration Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user with email
      tags:
      - Auth
  /api/auth/email/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using token
      parameters:
      - description: Password reset token
        in: query
        name: token
        required: true
        type: string
      - description: New Password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Auth
  /api/auth/email/verify-email:
    get:
      description: Verify email using token
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirects to front-end after verification
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify user email
      tags:
      - Auth
  /api/auth/google/callback:
    get:
      description: Complete authentication after Google login
      produces:
      - application/json
      responses:
        "302":
          description: Redirects to front-end with token
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Google OAuth2 callback
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"